Improve SingleChart styling
And refactor syntax highlighting component out from FunctionChart
