+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./timeRanges";
 export * from "./uniq";
 export * from "./uniqBy";
 export * from "./getNonce";
+export * from "./getTitle";
